feat: added new Sequence extensions
