+++ Utils/README.md
@@ -1,3 +1,7 @@
 # Utils
 
 A description of this package.
+
+## Credits
+
+Sequence extensions: https://www.swiftbysundell.com/articles/async-and-concurrent-forEach-and-map/
