--- Utils/Sources/Utils/AsyncMap.swift
@@ -1,22 +1,0 @@
-//
-//  AsyncMap.swift
-//  
-//
-//  Created by Егор Яковенко on 26.02.2022.
-//
-
-import Foundation
-
-public extension Sequence {
-    func asyncMap<T>(
-        _ transform: (Element) async throws -> T
-    ) async rethrows -> [T] {
-        var values = [T]()
-
-        for element in self {
-            try await values.append(transform(element))
-        }
-
-        return values
-    }
-}
