+++ Utils/Sources/Utils/Sequence/Sequence+ForEach.swift
@@ -1,0 +1,33 @@
+//
+//  Sequence+ForEach.swift
+//  
+//
+//  Created by Егор Яковенко on 27.02.2022.
+//
+
+extension Sequence {
+    func asyncForEach(
+        _ operation: (Element) async throws -> Void
+    ) async rethrows {
+        for element in self {
+            try await operation(element)
+        }
+    }
+}
+
+extension Sequence {
+    func concurrentForEach(
+        _ operation: @escaping (Element) async -> Void
+    ) async {
+        // A task group automatically waits for all of its
+        // sub-tasks to complete, while also performing those
+        // tasks in parallel:
+        await withTaskGroup(of: Void.self) { group in
+            for element in self {
+                group.addTask {
+                    await operation(element)
+                }
+            }
+        }
+    }
+}
