+++ Utils/Sources/Utils/Sequence/Sequence+Map.swift
@@ -1,0 +1,38 @@
+//
+//  Sequence+Map.swift
+//  
+//
+//  Created by Егор Яковенко on 26.02.2022.
+//
+
+import Foundation
+
+public extension Sequence {
+    func asyncMap<T>(
+        _ transform: (Element) async throws -> T
+    ) async rethrows -> [T] {
+        var values = [T]()
+
+        for element in self {
+            try await values.append(transform(element))
+        }
+
+        return values
+    }
+}
+
+extension Sequence {
+    func concurrentMap<T>(
+        _ transform: @escaping (Element) async throws -> T
+    ) async throws -> [T] {
+        let tasks = map { element in
+            Task {
+                try await transform(element)
+            }
+        }
+        
+        return try await tasks.asyncMap { task in
+            try await task.value
+        }
+    }
+}
